Correction des unités du moteur physique + trajectoires
passage d’un système en pixel/ms à pixel/s

+++ src/eu/labrush/rescue/core/physics/AbstractPhysicBehaviour.java
@@ -36,7 +36,7 @@ public abstract class AbstractPhysicBehaviour {
 		this.dim = dim;
 	}
 
-	public abstract void updateTrajectoire(HashSet<? extends AbstractObject> obstacles, int delta_t);
+	public abstract void updateTrajectoire(HashSet<? extends AbstractObject> obstacles, double delta_t);
 
 	/**
 	 * @param bloc le bloc par rapport auquel on se situe
