+++ src/eu/labrush/rescue/core/physics/RebondPhysicBehaviour.java
@@ -31,7 +31,7 @@ public class RebondPhysicBehaviour extends AbstractPhysicBehaviour {
 	 * int)
 	 */
 	@Override
-	public void updateTrajectoire(HashSet<? extends AbstractObject> obstacles, int delta_t) {
+	public void updateTrajectoire(HashSet<? extends AbstractObject> obstacles, double delta_t) {
 		
 		this.moves = new LibertyDegree();
 		Vecteur deplacement = trajectoire.distanceParcourue(delta_t);
